Support simple export interface
Remotes that support `exporttree=yes` and `git annex export can now be
implemented using this library. See [1] for details of this part of the
protocol.

According to the git-annex documentation, the methods that are included
here as part of the export interface can technically be supported or not
on an individual basis. However, an export remote is completely useless
without at least CHECKPRESENTEXPORT and TRANSFEREXPORT STORE;
TRANSFEREXPORT RETRIEVE and REMOVEEXPORT are not quite as critical, but
they're still important and not supporting them can cause basic remote
operations to fail. So we include all of them in a single interface to
simplify the library and encourage fully functional remotes. If any of
them really can't be implemented for some reason, the function can
return an error.

REMOVEEXPORTDIRECTORY and RENAMEEXPORT are much more peripheral and
explicitly optional (not supporting them won't cause operations to
fail), so they remain separate.

[1] https://git-annex.branchable.com/design/external_special_remote_protocol/export_and_import_appendix/