Move remote code into a subdirectory
There are other line-based protocols that git-annex uses similar to the
one for external special remotes -- the external backend protocol and
the P2P protocol -- and it could make sense to support those in this
module as well. (Well, at least the backend one; the P2P protocol is
documented, but I'm not sure if it's useful to write external tools that
speak it.) This change just moves everything wholesale for now, but
ideally we'll end up with common code in one place and protocol-specific
code in subdirectories.